app/models/api_key.py
from app.extensions import db
from datetime import datetime
from sqlalchemy.dialects.mysql import JSON

class APIKey(db.Model):
    __tablename__ = 'api_keys'

    id = db.Column(db.Integer, primary_key=True)
    key_hash = db.Column(db.String(64), unique=True, nullable=False)
    name = db.Column(db.String(50), nullable=False)
    scopes = db.Column(JSON)
    is_active = db.Column(db.Boolean, default=True)
    created_at = db.Column(db.DateTime, default=datetime.utcnow)
    expires_at = db.Column(db.DateTime, nullable=True)
    last_used_at = db.Column(db.DateTime, nullable=True)
    created_by = db.Column(db.String(50), nullable=True)

    def __repr__(self):
        return f'<APIKey {self.name}>' 