requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-limiter==3.5.0
python-dotenv==1.0.0
requests==2.31.0
mysql-connector-python==8.2.0
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1
pandas==2.1.4
numpy==1.26.2
pytest==8.0.2
responses==0.24.1
pytest-mock==3.12.0 